fix response api

+++ src/main/java/com/project/soshuceapi/controllers/PetController.java
@@ -11,7 +11,9 @@ import com.project.soshuceapi.models.requests.PetUpdateRequest;
 import com.project.soshuceapi.models.responses.Error;
 import com.project.soshuceapi.models.responses.Response;
 import com.project.soshuceapi.services.PetService;
+import com.project.soshuceapi.utils.StringUtil;
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotBlank;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.http.HttpStatus;
@@ -51,35 +53,45 @@ public class PetController {
             @RequestParam(value = "sterilization", defaultValue = "", required = false) Integer sterilization,
             @RequestParam(value = "rabies", defaultValue = "", required = false) Integer rabies
     ) {
+        Response<Map<String, Object>> response = new Response<>();
+        response.setSuccess(false);
         try {
-            Response<Map<String, Object>> response = new Response<>();
             response.setData(petService.getAll(page, limit,
                     name, breed, color, code, type, age, gender,
                     status, diet, vaccin, sterilization, rabies));
             response.setSuccess(true);
             return ResponseEntity.ok(response);
         } catch (Exception e) {
-            return ResponseEntity.badRequest().body(e.getMessage());
+            response.setError(Error.of(e.getMessage(), ResponseCode.Common.FAIL));
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
         }
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<?> getPetById(@PathVariable("id") String id) {
+        Response<PetDTO> response = new Response<>();
+        response.setSuccess(false);
         try {
-            Response<PetDTO> response = new Response<>();
+            if (StringUtil.isNullOrBlank(id)) {
+                response.setError(Error.of(ResponseMessage.Common.INVALID_INPUT, ResponseCode.Common.INVALID));
+                return ResponseEntity.badRequest().body(response);
+            }
             response.setData(petService.getById(id));
             response.setSuccess(true);
             return ResponseEntity.ok(response);
         } catch (NotFoundException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+            response.setError(Error.of(ResponseMessage.Common.NOT_FOUND, ResponseCode.Common.NOT_FOUND));
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
         } catch (Exception e) {
-            return ResponseEntity.badRequest().body(e.getMessage());
+            response.setError(Error.of(e.getMessage(), ResponseCode.Common.FAIL));
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
         }
     }
 
     @GetMapping("/statistic-cases")
     public ResponseEntity<?> getStatisticStatus() {
         Response<Map<String, Long>> response = new Response<>();
+        response.setSuccess(false);
         try {
             response.setData(petService.getStatisticCases());
             response.setSuccess(true);
@@ -94,14 +106,16 @@ public class PetController {
     @PreAuthorize("hasRole('ADMIN') || hasRole('MANAGER')")
     public ResponseEntity<?> create(@Valid @ModelAttribute PetCreateRequest request, BindingResult bindingResult) {
         Response<String> response = new Response<>();
+        response.setSuccess(false);
         try {
-            if (bindingResult.hasErrors()) {
-                response.setError(Error.of(Objects.requireNonNull(bindingResult.getFieldError()).getDefaultMessage(), ResponseCode.Common.FAIL));
-                return ResponseEntity.badRequest().body(response);
-            }
             if (auditorAware.getCurrentAuditor().isEmpty()) {
                 response.setError(Error.of(ResponseMessage.Authentication.PERMISSION_DENIED, ResponseCode.Authentication.PERMISSION_DENIED));
                 return ResponseEntity.status(HttpStatus.FORBIDDEN).body(response);
+            }
+            if (bindingResult.hasErrors()) {
+                response.setError(Error.of(Objects.requireNonNull(bindingResult.getFieldError()).getDefaultMessage(),
+                        ResponseCode.Common.INVALID));
+                return ResponseEntity.badRequest().body(response);
             }
             request.setCreatedBy(auditorAware.getCurrentAuditor().get());
             response.setData(petService.create(request).getId());
@@ -116,20 +130,22 @@ public class PetController {
     @PutMapping("/update/{id}")
     @PreAuthorize("hasRole('ADMIN') || hasRole('MANAGER')")
     public ResponseEntity<?> update(@Valid @ModelAttribute PetUpdateRequest request,
-                                    @RequestParam(name = "id") String id,
+                                    @PathVariable("id") String id,
                                     BindingResult bindingResult) {
         Response<String> response = new Response<>();
+        response.setSuccess(false);
         try {
-            if (bindingResult.hasErrors()) {
-                response.setError(Error.of(Objects.requireNonNull(bindingResult.getFieldError()).getDefaultMessage(), ResponseCode.Common.FAIL));
-                return ResponseEntity.badRequest().body(response);
-            }
             if (auditorAware.getCurrentAuditor().isEmpty()) {
                 response.setError(Error.of(ResponseMessage.Authentication.PERMISSION_DENIED, ResponseCode.Authentication.PERMISSION_DENIED));
                 return ResponseEntity.status(HttpStatus.FORBIDDEN).body(response);
             }
+            if (bindingResult.hasErrors()) {
+                response.setError(Error.of(Objects.requireNonNull(bindingResult.getFieldError()).getDefaultMessage(),
+                        ResponseCode.Common.INVALID));
+                return ResponseEntity.badRequest().body(response);
+            }
             if (!id.equals(request.getId())) {
-                response.setError(Error.of(ResponseMessage.Common.NOT_MATCH, ResponseCode.Common.FAIL));
+                response.setError(Error.of(ResponseMessage.Common.NOT_MATCH, ResponseCode.Common.NOT_MATCH));
                 return ResponseEntity.badRequest().body(response);
             }
             request.setUpdatedBy(auditorAware.getCurrentAuditor().get());
@@ -137,32 +153,42 @@ public class PetController {
             response.setSuccess(true);
             return ResponseEntity.ok(response);
         } catch (NotFoundException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+            response.setError(Error.of(ResponseMessage.Common.NOT_FOUND, ResponseCode.Common.NOT_FOUND));
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
         } catch (Exception e) {
             response.setError(Error.of(e.getMessage(), ResponseCode.Common.FAIL));
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
         }
     }
 
-    @PutMapping("/update-image")
+    @PutMapping("/update-image/{id}")
     @PreAuthorize("hasRole('ADMIN') || hasRole('MANAGER')")
-    public ResponseEntity<?> updateImage(@Valid @ModelAttribute PetUpdateImageRequest request, BindingResult bindingResult) {
+    public ResponseEntity<?> updateImage(@Valid @ModelAttribute PetUpdateImageRequest request,
+                                         @PathVariable("id") String id,
+                                         BindingResult bindingResult) {
         Response<String> response = new Response<>();
+        response.setSuccess(false);
         try {
-            if (bindingResult.hasErrors()) {
-                response.setError(Error.of(Objects.requireNonNull(bindingResult.getFieldError()).getDefaultMessage(), ResponseCode.Common.FAIL));
-                return ResponseEntity.badRequest().body(response);
-            }
             if (auditorAware.getCurrentAuditor().isEmpty()) {
                 response.setError(Error.of(ResponseMessage.Authentication.PERMISSION_DENIED, ResponseCode.Authentication.PERMISSION_DENIED));
                 return ResponseEntity.status(HttpStatus.FORBIDDEN).body(response);
+            }
+            if (bindingResult.hasErrors()) {
+                response.setError(Error.of(Objects.requireNonNull(bindingResult.getFieldError()).getDefaultMessage(),
+                        ResponseCode.Common.INVALID));
+                return ResponseEntity.badRequest().body(response);
+            }
+            if (!id.equals(request.getId())) {
+                response.setError(Error.of(ResponseMessage.Common.NOT_MATCH, ResponseCode.Common.NOT_MATCH));
+                return ResponseEntity.badRequest().body(response);
             }
             request.setUpdatedBy(auditorAware.getCurrentAuditor().get());
             response.setData(petService.updateImage(request).getId());
             response.setSuccess(true);
             return ResponseEntity.ok(response);
         } catch (NotFoundException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+            response.setError(Error.of(ResponseMessage.Common.NOT_FOUND, ResponseCode.Common.NOT_FOUND));
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
         } catch (Exception e) {
             response.setError(Error.of(e.getMessage(), ResponseCode.Common.FAIL));
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
@@ -173,15 +199,21 @@ public class PetController {
     @PreAuthorize("hasRole('ADMIN') || hasRole('MANAGER')")
     public ResponseEntity<?> deleteSoft(@PathVariable("id") String id) {
         Response<String> response = new Response<>();
+        response.setSuccess(false);
         try {
             if (auditorAware.getCurrentAuditor().isEmpty()) {
                 response.setError(Error.of(ResponseMessage.Authentication.PERMISSION_DENIED, ResponseCode.Authentication.PERMISSION_DENIED));
                 return ResponseEntity.status(HttpStatus.FORBIDDEN).body(response);
             }
+            if (StringUtil.isNullOrBlank(id)) {
+                response.setError(Error.of(ResponseMessage.Common.INVALID_INPUT, ResponseCode.Common.INVALID));
+                return ResponseEntity.badRequest().body(response);
+            }
             response.setSuccess(petService.deleteSoft(id));
             return ResponseEntity.ok(response);
         } catch (NotFoundException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+            response.setError(Error.of(ResponseMessage.Common.NOT_FOUND, ResponseCode.Common.NOT_FOUND));
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
         } catch (Exception e) {
             response.setError(Error.of(e.getMessage(), ResponseCode.Common.FAIL));
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
